Don't overwrite config/master.key even on --force
See https://github.com/rails/rails/pull/31957#issuecomment-364817423

The purpose of `--force` is not to have any prompt whether a file
should
be kept or overwritten. In general, all existing files should be
overwritten.
However, `config/master.key` is special because it is git-ignored, and
overwriting it will cause the app not to run (since there won't be a
way
to decrypt the credentials).

As a result, it's probably better to keep the existing
config/master.key.